updating signUpForm.js file on registration branch

+++ app/forms/venues_forms.py
@@ -1,5 +1,5 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, FloatField, IntegerField
+from wtforms import StringField, IntegerField
 from wtforms.validators import DataRequired, Length
 
 class VenueForm(FlaskForm):
@@ -9,8 +9,8 @@ class VenueForm(FlaskForm):
    city = StringField("city", validators=[DataRequired()])
    state = StringField("state", validators=[DataRequired()])
    zip_code = StringField("zip", validators=[DataRequired(), Length(max=5)])
-   latitude = FloatField("lat")
-   longitude = FloatField("long")
+   latitude = IntegerField("lat")
+   longitude = IntegerField("long")
 
    def to_dict(self):
       return{
